+++ game/missiongenerator/tgogenerator.py
@@ -446,7 +446,7 @@ class GenericCarrierGenerator(GroundObjectGenerator):
                 icls = next(self.icls_alloc)
                 link4 = None
                 if carrier_type in [Stennis, CVN_71, CVN_72, CVN_73, CVN_75]:
-                    link4 = self.radio_registry.alloc_uhf()
+                    link4 = self.radio_registry.alloc_link4()
                 self.activate_beacons(ship_group, tacan, tacan_callsign, icls, link4)
                 self.add_runway_data(
                     brc or Heading.from_degrees(0), atc, tacan, tacan_callsign, icls